requirements.txt
asgiref==3.2.7
braintree==4.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-import-export==2.0.2
django-payments==0.13.0
django-paypal==1.0.0
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
jdcal==1.4.1
lml==0.0.9
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
pyexcel-io==0.5.20
pyexcel-ods==0.5.6
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.45.0
suds-jurko==0.6
tablib==1.1.0
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
